Implement request for adding a community
Currently does not work because of the way libevent behaves. Large
files do not work because libevent has a max read of 2^14 bytes.
Sending back multiple communities does not work either because we
send them as the same "packet" which confuses the client. The
solution to this is adding a length field to the header
preferably uint64_t and implementing a system on both sides
that reads the data until it has enough data and then processes the
packet.

Made a seperate branch to keep master stable, client and server both
easy to crash at the moment if you feed them the right data.